+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Make `parse::parse` a method on `Expr`
+- Make `and` and `or` evaluation lazy
 
 
 ## [0.2.0] - 2022-11-27