+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.0] - 2022-11-29
 
 ### Added
 
@@ -76,7 +76,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement a `RuleSet` as an entry point for executing a set of rules on a piece of data 
 - async `UserFunction` support
 
-[unreleased]: https://github.com/mendelt/reval/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/mendelt/reval/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/mendelt/reval/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/mendelt/reval/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/mendelt/reval/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/mendelt/reval/compare/v0.1.1...v0.1.2